rename consumer wrapper to handler
